[FIX] Computed opening balance was wrong in some cases when there is opening entries in the very first opening period.
In almost all cases, the entries which are in the opening periods must not be considered for the computation of the
initial balances of partners. In one case, though they have to be considered: the opening entries in the very first year
in the accounting. Thoses entries do not have any counterpart the previous year and have to be summed to the initial balance.