[CHG] remove the configuration of the policy at the account level...
... add a configuration at the company level, applied on partners if they do not have a policy.
I did like to put a property but we cannot search on properties.

This follows the change on the policies, on which we now configure the active accounts for a policy.
This is more simpler for the user, and simplifies the code as well.

It also avoid the issue that we had when credit control lines could have been generated on accounts without policies
if the partner has a enforced policy.