[CHG] Introduced a new state "Ignored" and changed the behaviour with "Draft" lines. Details:
When a line, say, from the January run, is in state "Ignored" when the February run is launched,
the line will be generated again, without increasing the level, basically it will be the same line
than the January one, except its "controlling date".
When a line is left as "Draft", it is considered as "Ignored", so it will be generated again too,
and the first line will be put to "Ignored" state.
"Draft" and "Ignored" have the same behaviour, but not the same meaning for the end user,
it is a classifier.